Changed input to take in CSV from CrossAsymmetric substructure, still has not fully worked (Changing local directory)
Changed importSubstructure to take in the CSV file output from substructure generation. To do so there is now a function csvColRange that uses genfromtxt to get desired columns (substructure info rows portion) in the csv using a start column and end column for a range, from what I could see this was not built into the function.

The output of that function is set up similarly to the last iteration to match the input needed for nearestSubstructure.

Values are changed accordingly in each file to match the 1st optimization run of the CrossAsymmetric substructure. AssemblyModifyEdit.py still has not fully worked.

Moving local directory due to lack of space and pushing to GitHub in case I don't succesfully change the local directory.